The main bottleneck was the manipulation of datetimes to shift to the survey scheduled timezone, as well as formatting those datetimes, or repeatedly accessing `survey.schedule.timezone` when we could cache the values for the whole HTTP request.

That may not be the only bottleneck or optimization possible, but that's a first step that actually allows the CSV files to be generated faster. I don't have enough data to have a significant benchmark for the results and incentives CSV, and the disposition history CSV was already optimized, but the interactions CSV is now magnitudes faster:

- `results.csv` (60% faster):
  - main: 5s at 168 KB/s
  - patch: 2s at 431 KB/s

- `disposition_history.csv` (no improvement):
  - main: 2s at 1084 KB/s
  - patch: 2s at 1023 KB/s

- `incentives.csv` (66% faster):
  - main: 3s at 155 KB/s
  - patch: 1s 486 KB/s

- `interactions.csv` (93% faster):
  - main: 31s at 261 KB/s
  - patch: 2s at 1023 KB/s

@matiasgarciaisaia As far as I understand, SQL queries are negligible, the bottleneck is the Elixir/Erlang code that manipulates the data to generate the CSV files. The impact is _very_ noticeable.

I tried to replace the ugly serialization of DateTime to String, to use more readable functions now available in Elixir 1.10 such as `DateTime.add/4` or the legacy `Timex.format!/3` but they had too much of an impact (at least 2 or 3 times as slower). Newer versions of Elixir have a datetime formatter, maybe it will be faster, and it has been optimized since...

closes #2205