@ysbaddaden 2023-04-14
@leandroradusky I pushed a couple more commits to reduce the number of SQL requests, if you could give 'em a look :bow: