@matiasgarciaisaia 2023-04-15
Great! Can confirm 2.5x speedup, with same results, for interactions file 👍